requirements.txt
fastmcp
PyYAML
psycopg2-binary
pymongo
redis
pytest
